api/requirements.txt
fastapi
uvicorn
ib_insync
python-dotenv

